requirements.txt
httpx>=0.28
mistralai>=2.1
pdfplumber>=0.11
python-docx>=1.1
python-dotenv>=1.0
tiktoken>=0.12
tqdm>=4.66
